backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic[email]==2.6.4
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
slowapi==0.1.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
faker==24.3.0
python-dotenv==1.0.1
